% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPatchSeqDissocCellMarkerExpr.R
\name{plotPatchSeqDissocCellMarkerExpr}
\alias{plotPatchSeqDissocCellMarkerExpr}
\title{Plot the summarized marker expression for dissociated cell and patch seq data}
\usage{
plotPatchSeqDissocCellMarkerExpr(patchseqContam, dissocContam, on_type,
  off_type)
}
\arguments{
\item{patchseqContam}{contamination data frame for patch seq samples}

\item{dissocContam}{contamination data frame for dissociated cell samples}

\item{on_type}{string for which cell type is the on type - all patch seq samples should be of this type}

\item{off_type}{string for which cell type is the off type}
}
\value{
plot showing marker expression for both datasets
}
\description{
Plot the summarized marker expression for dissociated cell and patch seq data
}
